Move Synchronizer to :persistent-cache
