+++ platforms/core-execution/persistent-cache/src/main/java/org/gradle/cache/internal/CacheAccessSerializer.java
@@ -16,7 +16,6 @@
 package org.gradle.cache.internal;
 
 import org.gradle.cache.Cache;
-import org.gradle.internal.concurrent.Synchronizer;
 
 import java.util.function.Function;
